Moved UncheckedURISyntaxException to base level exception package
